Déplace les formulaires relatifs aux contributeurs avec leurs vues
