+++ zds/tutorialv2/views/contributors.py
@@ -1,5 +1,10 @@
 from collections import OrderedDict
 
+from crispy_forms.bootstrap import StrictButton
+from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Layout, Field
+from django import forms
+
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.models import User
@@ -12,15 +17,74 @@ from django.utils.text import format_lazy
 from django.utils.translation import gettext_lazy as _
 
 from zds.member.decorator import LoggedWithReadWriteHability
+from zds.member.models import Profile
 from zds.member.utils import get_bot_account
 from zds.notification.models import NewPublicationSubscription
 from zds.tutorialv2 import signals
-from zds.tutorialv2.forms import ContributionForm, RemoveContributionForm
+from zds.tutorialv2.forms import ReviewerTypeModelChoiceField
 from zds.tutorialv2.mixins import SingleContentFormViewMixin
 from zds.tutorialv2.models import TYPE_CHOICES_DICT
-from zds.tutorialv2.models.database import ContentContribution, PublishableContent
+from zds.tutorialv2.models.database import ContentContribution, PublishableContent, ContentContributionRole
 from zds.mp.utils import send_mp
 from zds.utils.paginator import ZdSPagingListView
+
+
+class ContributionForm(forms.Form):
+    contribution_role = ReviewerTypeModelChoiceField(
+        label=_("Role"),
+        required=True,
+        queryset=ContentContributionRole.objects.order_by("title").all(),
+    )
+
+    username = forms.CharField(
+        label=_("Contributeur"),
+        required=True,
+        widget=forms.TextInput(
+            attrs={"placeholder": _("Pseudo du membre à ajouter."), "data-autocomplete": "{ 'type': 'single' }"}
+        ),
+    )
+
+    comment = forms.CharField(
+        label=_("Commentaire"),
+        required=False,
+        widget=forms.Textarea(attrs={"placeholder": _("Commentaire sur ce contributeur."), "rows": "3"}),
+    )
+
+    def __init__(self, content, *args, **kwargs):
+        self.helper = FormHelper()
+        self.helper.form_class = "modal modal-flex"
+        self.helper.form_id = "add-contributor"
+        self.helper.form_method = "post"
+        self.helper.form_action = reverse("content:add-contributor", kwargs={"pk": content.pk})
+        self.helper.layout = Layout(
+            Field("username"),
+            Field("contribution_role"),
+            Field("comment"),
+            StrictButton(_("Ajouter"), type="submit", css_class="btn-submit"),
+        )
+        super().__init__(*args, **kwargs)
+
+    def clean_username(self):
+        cleaned_data = super().clean()
+        if cleaned_data.get("username"):
+            username = cleaned_data.get("username")
+            user = Profile.objects.contactable_members().filter(user__username__iexact=username.strip().lower()).first()
+            if user is not None:
+                cleaned_data["user"] = user.user
+            else:
+                self._errors["user"] = self.error_class([_("L'utilisateur sélectionné n'existe pas")])
+
+        if "user" not in cleaned_data:
+            self._errors["user"] = self.error_class([_("Veuillez renseigner l'utilisateur")])
+
+        return cleaned_data
+
+
+class RemoveContributionForm(forms.Form):
+    pk_contribution = forms.CharField(
+        label=_("Contributeur"),
+        required=True,
+    )
 
 
 class AddContributorToContent(LoggedWithReadWriteHability, SingleContentFormViewMixin):
